packages/pollination-honeybee-energy/pollination-honeybee-energy-0.2.10.tar.gz/pollination-honeybee-energy-0.2.10/tests/settings_test.py
from pollination.honeybee_energy.settings import SimParDefault, SimParLoadBalance, \
    SimParComfort, BaselineOrientationSimPars
from queenbee.plugin.function import Function


def test_sim_par_default():
    function = SimParDefault().queenbee
    assert function.name == 'sim-par-default'
    assert isinstance(function, Function)


def test_sim_par_load_balance():
    function = SimParLoadBalance().queenbee
    assert function.name == 'sim-par-load-balance'
    assert isinstance(function, Function)


def test_sim_par_comfort():
    function = SimParComfort().queenbee
    assert function.name == 'sim-par-comfort'
    assert isinstance(function, Function)


def test_baseline_orientation_sim_pars():
    function = BaselineOrientationSimPars().queenbee
    assert function.name == 'baseline-orientation-sim-pars'
    assert isinstance(function, Function)
